feat: change shift_length computation method
The new method considers the availability of the lane on the left side or right
side of ego vehicle, depending on the position of the obstacles.

Will not generate avoidance path if the adjacent lane is not a lanelet.

this is a part of .iv PR (tier4/autoware.iv#2383) port

Signed-off-by: Muhammad Zulfaqar Azmi <zulfaqar.azmi@tier4.jp>