refactor(lidar_apollo_segmentation_tvm): rework parameters (#3390)
Remove duplication of parameter declaration between the source code and
the parameter file.

Remove the default value from the source code in order to ensure all
parameter values are passed from the parameter files.

Use types that match the ones returned from declare_parameter() in order
to prevent confusion and potential coding errors.

Move the default parameter file to a standardized location. Use the
JSON Schema for that.

Issue-Id: SCM-5887

Change-Id: Idc2774c4821334ae93027bf4baaa0e543aeebe3a

Signed-off-by: Ambroise Vincent <ambroise.vincent@arm.com>
Co-authored-by: Xinyu Wang <93699235+angry-crab@users.noreply.github.com>
Signed-off-by: Mingyu Li <mingyu.li@tier4.jp>