Optimize memory usage of walks array
The walks array in the previous implementation could in the worst case grow
to the number of input points (if every point is an outlier).
This increased computation time for gradually growing the walks vector.

The current implementation only saves the first and current walk for
every ring. The isCluster check is performed right when a walk is
completed, and a bool vector akin to the previous walk_is_cluster
vector is maintained.

Signed-off-by: Maximilian Schmeller <maximilian.schmeller@tier4.jp>