Check for invalid pointcloud format
In the previous revision (and before this PR), pointcloud field indices
in ring_outlier_filter were hardcoded and fields were assumed to exist.

VRichardJP solved this by implementing individual accessor functions,
which ended up to be a bit verbose for this small piece of
functionality.

Now, there is one generic accessor function and the faster_filter
function exits if any of the fields expected does not exist.
The expected field data type is also checked.

Signed-off-by: Maximilian Schmeller <maximilian.schmeller@tier4.jp>