feat(behavior_path_planner): Add option to insert zero velocity to the center line path
This commit adds a new boolean parameter, `insert_zero_velocity`, to the `getCenterLinePath` function in the `utils.hpp` file. When set to `true`, this parameter will insert a zero velocity to each point in the path. In the `utils.cpp` file, this option is used to insert zero velocities to the reference path generated in the `createGoalAroundPath` function.

Signed-off-by: kyoichi-sugahara <kyoichi.sugahara@tier4.jp>